NNE - Warning Push

+++ Game/Source/Application.h
@@ -99,7 +99,7 @@ private: // JSON Functions
 	{
 		float array[3] = {};
 		for (int i = 0; i < 3; i++)
-			array[i] = json_array_get_number(arr, (i + (3 * offSet)));
+			array[i] = float(json_array_get_number(arr, (i + (3 * offSet))));
 
 		vec->x = array[0];
 		vec->y = array[1];
@@ -114,7 +114,7 @@ private: // JSON Functions
 		// If it is a "AllVectorArray", you can inout the offset thinking about vectors as units, not a group of 3 units.
 		// So you can input which vector you want inside this "AllVectorArray". 'Vector 1, Vector 2, Vector 3' instead of 'Vector(unit 1, unit 2, unit 3), Vector(unit 4, unit 5, unit, 6)...'
 
-		for (int i = 0; i < 3; i++) array[i] = json_array_get_number(arr, (i + (mult * offSet)));
+		for (unsigned int i = 0; i < 3; i++) array[i] = float(json_array_get_number(arr, (i + (mult * offSet))));
 
 		vec->x = array[0];
 		vec->y = array[1];
